market_server/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-xml==2.0.0
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
packaging==23.1
protobuf==3.20.3
psycopg==3.1.9
pytz==2023.3
requests==2.31.0
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
